added search test
