+++ src/test/java/Tests/brain/gadgets/CheckDronesCategory.java
@@ -1,15 +1,18 @@
-package Tests.brain;
+package Tests.brain.gadgets;
 
 import PageObject.BasePage;
-import PageObject.brain.com.ua.*;
+import PageObject.brain.com.ua.BrainGadgetsPage;
+import PageObject.brain.com.ua.BrainHomePage;
+import PageObject.brain.com.ua.BrainOrderPage;
+import PageObject.brain.com.ua.BrainProductPage;
 import Tests.TestInit;
 import org.testng.Assert;
 import org.testng.annotations.Test;
 
-public class CheckCamerasCategory extends TestInit {
+public class CheckDronesCategory extends TestInit {
 
     @Test
-    public void checkCamerasCategory() {
+    public void checkDronesCategory() {
 
         goToSite("https://brain.com.ua/ukr/");
         BasePage page = new BasePage(driver);
@@ -19,7 +22,8 @@ public class CheckCamerasCategory extends TestInit {
         homePage.getGadgets().click();
 
         BrainGadgetsPage gadgetsPage = new BrainGadgetsPage(driver);
-        gadgetsPage.getDigitalCameras().click();
+        gadgetsPage.getDrones().click();
+        gadgetsPage.getSubPropellers().click();
         homePage.getFirstModel().click();
 
         BrainProductPage productPage = new BrainProductPage(driver);
